\name{association-package}
\alias{association-package}
\alias{association}
\docType{package}
\title{
\packageTitle{association}
}
\description{
\packageDescription{association}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{association}
\packageIndices{association}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{association}

Maintainer: \packageMaintainer{association}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
